requirements.txt
pandas==2.1.1
apache-airflow==2.7.1
pydantic-settings==2.0.3
